Restructures component directories
