+++ src/Common/ArrowSelector/ArrowSelector.css
